refactor(oc-path): share JSONC positional resolver
